Add InventoryIncrementalJobPublicAPI, move PipelineJobPublicAPI processConfig apis
