+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/api/MigrationJobPublicAPI.java
@@ -36,7 +36,7 @@ import java.util.Properties;
  * Migration job public API.
  */
 @SingletonSPI
-public interface MigrationJobPublicAPI extends PipelineJobPublicAPI, RequiredSPI {
+public interface MigrationJobPublicAPI extends InventoryIncrementalJobPublicAPI, RequiredSPI {
     
     /**
      * List all jobs.
