+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/api/PipelineJobPublicAPIFactory.java
@@ -28,18 +28,18 @@ import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPIRegistry;
 public final class PipelineJobPublicAPIFactory {
     
     static {
-        ShardingSphereServiceLoader.register(PipelineJobPublicAPI.class);
+        ShardingSphereServiceLoader.register(InventoryIncrementalJobPublicAPI.class);
         ShardingSphereServiceLoader.register(MigrationJobPublicAPI.class);
     }
     
     /**
-     * Get instance of pipeline job public API.
+     * Get instance of inventory incremental job public API.
      *
      * @param jobTypeName job type name
      * @return got instance
      */
-    public static PipelineJobPublicAPI getPipelineJobPublicAPI(@NonNull final String jobTypeName) {
-        return TypedSPIRegistry.getRegisteredService(PipelineJobPublicAPI.class, jobTypeName);
+    public static InventoryIncrementalJobPublicAPI getInventoryIncrementalJobPublicAPI(@NonNull final String jobTypeName) {
+        return TypedSPIRegistry.getRegisteredService(InventoryIncrementalJobPublicAPI.class, jobTypeName);
     }
     
     /**
